Rename interfaces and features
