+++ src/Filesystem/Bridge/Symfony/DependencyInjection/ZenstruckFilesystemExtension.php
@@ -32,8 +32,8 @@ use Zenstruck\Filesystem\Bridge\Symfony\Routing\RouteFileUrlFeature;
 use Zenstruck\Filesystem\Bridge\Twig\ObjectNodeLoaderExtension;
 use Zenstruck\Filesystem\Feature\FileUrl;
 use Zenstruck\Filesystem\Feature\FileUrl\PrefixFileUrlFeature;
-use Zenstruck\Filesystem\Feature\GlideUrl;
-use Zenstruck\Filesystem\Feature\GlideUrl\UrlBuilderGlideUrlFeature;
+use Zenstruck\Filesystem\Feature\TransformImageUrl;
+use Zenstruck\Filesystem\Feature\TransformImageUrl\GlideTransformImageUrl;
 use Zenstruck\Filesystem\LoggableFilesystem;
 use Zenstruck\Filesystem\MultiFilesystem;
 use Zenstruck\Filesystem\ReadonlyFilesystem;
@@ -205,11 +205,11 @@ final class ZenstruckFilesystemExtension extends ConfigurableExtension
         }
 
         if ($config['glide_url_builder']) {
-            $container->register($id = '.zenstruck_filesystem.feature.'.$name.'_glide_url_builder', UrlBuilderGlideUrlFeature::class)
+            $container->register($id = '.zenstruck_filesystem.feature.'.$name.'_glide_url_builder', GlideTransformImageUrl::class)
                       ->setArguments([new Reference($config['glide_url_builder'])])
             ;
 
-            $features[GlideUrl::class] = new Reference($id);
+            $features[TransformImageUrl::class] = new Reference($id);
         }
 
         if ($config['route']['name']) {
